Move reusable functions from transmuxer tests to utils module
